Allow extending extensions without a full mode
