+++ lib/core/src/server/config.js
@@ -4,6 +4,7 @@ import { createDefaultWebpackConfig } from './config/defaults/webpack.config';
 import devBabelConfig from './config/babel';
 import loadCustomBabelConfig from './loadCustomBabelConfig';
 import loadCustomWebpackConfig from './loadCustomWebpackConfig';
+import mergeConfigs from './mergeConfigs';
 
 const noopWrapper = config => config;
 
@@ -15,36 +16,6 @@ function getBabelConfig({
 }) {
   const defaultConfig = wrapDefaultBabelConfig(defaultBabelConfig);
   return wrapBabelConfig(loadCustomBabelConfig(configDir, defaultConfig));
-}
-
-function mergeConfigs(config, customConfig) {
-  return {
-    ...customConfig,
-    // We'll always load our configurations after the custom config.
-    // So, we'll always load the stuff we need.
-    ...config,
-    // Override with custom devtool if provided
-    devtool: customConfig.devtool || config.devtool,
-    // We need to use our and custom plugins.
-    plugins: [...config.plugins, ...(customConfig.plugins || [])],
-    module: {
-      ...config.module,
-      // We need to use our and custom rules.
-      ...customConfig.module,
-      rules: [
-        ...config.module.rules,
-        ...((customConfig.module && customConfig.module.rules) || []),
-      ],
-    },
-    resolve: {
-      ...config.resolve,
-      ...customConfig.resolve,
-      alias: {
-        ...config.alias,
-        ...(customConfig.resolve && customConfig.resolve.alias),
-      },
-    },
-  };
 }
 
 function informAboutCustomConfig(defaultConfigName) {
